# docs/plans/api-specification.yaml
openapi: 3.1.0
info:
  title: ekaiX AIXcelerator API
  version: 1.0.0
  description: |
    Backend API for ekaiX — a Snowflake Native App that enables business users to create
    semantic models and Cortex Agents through AI conversation.

    **Authentication:** All requests are authenticated via the `Sf-Context-Current-User` header
    injected by SPCS (Snowpark Container Services). No separate auth system exists.

    **Workspace Isolation:** All data is scoped to the authenticated user's workspace via
    PostgreSQL Row-Level Security. Sharing is explicit via the shares endpoints.

    **Streaming:** Agent responses are delivered via Server-Sent Events (SSE) on the
    `/agent/stream/{session_id}` endpoint.
  contact:
    name: ekai Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://{spcs_endpoint}
    description: SPCS production
    variables:
      spcs_endpoint:
        default: ekaix.snowflakecomputing.app
...
security:
  - spcsAuth: []
components:
  securitySchemes:
    spcsAuth:
      type: apiKey
      in: header
      name: Sf-Context-Current-User
      description: Snowflake username injected by SPCS. Cannot be spoofed in production.

  # ==========================================================================
  # SCHEMAS
  # ==========================================================================
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error code (e.g., NOT_FOUND, VALIDATION_ERROR, FORBIDDEN)
          examples: ["NOT_FOUND"]
        message:
          type: string
          description: Human-readable error description
          examples: ["Data product not found"]
        details:
          type: object
          description: Additional error context (validation errors, field-level issues)
          additionalProperties: true
...
    Artifact:
      type: object
      required: [id, data_product_id, artifact_type, created_at]
      properties:
        id:
          type: string
          format: uuid
        data_product_id:
          type: string
          format: uuid
        artifact_type:
          type: string
          enum: [erd, yaml, brd, document, export]
        filename:
          type: string
        file_size_bytes:
          type: integer
        content_type:
          type: string
        metadata:
          type: object
          additionalProperties: true
        download_url:
          type: string
          format: uri
          description: Pre-signed MinIO URL (valid for 1 hour)
        created_at:
          type: string
          format: date-time

    # --- Documents ---
...
paths:
...
  /artifacts/{id}:
    get:
      tags: [Artifacts]
      operationId: getArtifact
      summary: Get artifact details and download URL
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Artifact details with pre-signed download URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags: [Artifacts]
      operationId: updateArtifact
      summary: Update artifact metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Updated artifact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
...